pyproject.toml
[project]
name = "transcripts"
version = "0.1.0"
description = "Generate transcripts from mediasite videos"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python~=0.2.0",
    "openai-whisper==20240930",
    "yt-dlp==2024.09.27",
    "PyQt6",
]

